+++ olex/modes/move.h
@@ -13,7 +13,7 @@ public:
 
     TXAtomPList Atoms;
     TGlXApp::GetGXApp()->FindXAtoms(Cmds.Text(' '), Atoms, true);
-    for( int i=0; i < Atoms.Count(); i++ )
+    for( size_t i=0; i < Atoms.Count(); i++ )
       Center += Atoms[i]->Atom().ccrd();
 
     if( Atoms.Count() != 0 )  Center /= Atoms.Count();
