+++ sdl/macroerror.cpp
@@ -33,7 +33,7 @@ void TMacroError::NonexitingMacroError(const olxstr& macroName)  {
   ProcessError |= peNonexistingFunction;
 }
 //..............................................................................
-void TMacroError::WrongArgCount(const ABasicFunction& func, unsigned int ArgC)  {
+void TMacroError::WrongArgCount(const ABasicFunction& func, size_t ArgC)  {
   ErrorInfo = "Macro/function '";
   ErrorInfo << func.GetSignature() << "' is provided with " << (int)ArgC << " arguments";
   ProcessError |= peInvalidArgCount;
